Link C and C++ API tests against the API-restricted shared library
