+++ CMakeLists.txt
@@ -8,10 +8,8 @@ set(CMAKE_FIND_PACKAGE_RESOLVE_SYMLINKS TRUE)
 set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
-if (MSVC)
-    set(CMAKE_CXX_VISIBILITY_PRESET hidden)
-    set(CMAKE_VISIBILITY_INLINES_HIDDEN ON)
-endif()
+set(CMAKE_CXX_VISIBILITY_PRESET hidden)
+set(CMAKE_VISIBILITY_INLINES_HIDDEN ON)
 
 # Detect OS and architecture, copied from DuckDB
 set(OS_NAME "unknown")
@@ -92,6 +90,9 @@ if(MSVC)
     # Remove the default to avoid warnings
     STRING(REPLACE "/EHsc" "" CMAKE_CXX_FLAGS ${CMAKE_CXX_FLAGS})
     STRING(REPLACE "/EHs" "" CMAKE_CXX_FLAGS ${CMAKE_CXX_FLAGS})
+    # Store all libraries and binaries in the same directory so that kuzu_shared.dll is found at runtime
+    set(LIBRARY_OUTPUT_PATH "${CMAKE_BINARY_DIR}/src")
+    set(EXECUTABLE_OUTPUT_PATH "${CMAKE_BINARY_DIR}/src")
 endif()
 if(CMAKE_BUILD_TYPE MATCHES Release)
     if(MSVC)
@@ -147,6 +148,15 @@ function(add_kuzu_test TEST_NAME)
     set(SRCS ${ARGN})
     add_executable(${TEST_NAME} ${SRCS})
     target_link_libraries(${TEST_NAME} PRIVATE test_helper test_runner graph_test)
+    target_include_directories(${TEST_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/test/include)
+    include(GoogleTest)
+    gtest_discover_tests(${TEST_NAME})
+endfunction()
+
+function(add_kuzu_api_test TEST_NAME)
+    set(SRCS ${ARGN})
+    add_executable(${TEST_NAME} ${SRCS})
+    target_link_libraries(${TEST_NAME} PRIVATE api_graph_test api_test_helper)
     target_include_directories(${TEST_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/test/include)
     include(GoogleTest)
     gtest_discover_tests(${TEST_NAME})
