+++ src/common/vector/value_vector.cpp
@@ -372,22 +372,22 @@ void ArrowColumnVector::slice(ValueVector* vector, offset_t offset) {
     setArrowColumn(vector, arrowColumnBuffer->column->Slice((int64_t)offset));
 }
 
-template void ValueVector::setValue<nodeID_t>(uint32_t pos, nodeID_t val);
-template void ValueVector::setValue<bool>(uint32_t pos, bool val);
-template void ValueVector::setValue<int64_t>(uint32_t pos, int64_t val);
-template void ValueVector::setValue<int32_t>(uint32_t pos, int32_t val);
-template void ValueVector::setValue<int16_t>(uint32_t pos, int16_t val);
-template void ValueVector::setValue<int8_t>(uint32_t pos, int8_t val);
-template void ValueVector::setValue<uint64_t>(uint32_t pos, uint64_t val);
-template void ValueVector::setValue<uint32_t>(uint32_t pos, uint32_t val);
-template void ValueVector::setValue<uint16_t>(uint32_t pos, uint16_t val);
-template void ValueVector::setValue<uint8_t>(uint32_t pos, uint8_t val);
-template void ValueVector::setValue<double_t>(uint32_t pos, double_t val);
-template void ValueVector::setValue<float_t>(uint32_t pos, float_t val);
-template void ValueVector::setValue<date_t>(uint32_t pos, date_t val);
-template void ValueVector::setValue<timestamp_t>(uint32_t pos, timestamp_t val);
-template void ValueVector::setValue<interval_t>(uint32_t pos, interval_t val);
-template void ValueVector::setValue<list_entry_t>(uint32_t pos, list_entry_t val);
+template KUZU_API void ValueVector::setValue<nodeID_t>(uint32_t pos, nodeID_t val);
+template KUZU_API void ValueVector::setValue<bool>(uint32_t pos, bool val);
+template KUZU_API void ValueVector::setValue<int64_t>(uint32_t pos, int64_t val);
+template KUZU_API void ValueVector::setValue<int32_t>(uint32_t pos, int32_t val);
+template KUZU_API void ValueVector::setValue<int16_t>(uint32_t pos, int16_t val);
+template KUZU_API void ValueVector::setValue<int8_t>(uint32_t pos, int8_t val);
+template KUZU_API void ValueVector::setValue<uint64_t>(uint32_t pos, uint64_t val);
+template KUZU_API void ValueVector::setValue<uint32_t>(uint32_t pos, uint32_t val);
+template KUZU_API void ValueVector::setValue<uint16_t>(uint32_t pos, uint16_t val);
+template KUZU_API void ValueVector::setValue<uint8_t>(uint32_t pos, uint8_t val);
+template KUZU_API void ValueVector::setValue<double_t>(uint32_t pos, double_t val);
+template KUZU_API void ValueVector::setValue<float_t>(uint32_t pos, float_t val);
+template KUZU_API void ValueVector::setValue<date_t>(uint32_t pos, date_t val);
+template KUZU_API void ValueVector::setValue<timestamp_t>(uint32_t pos, timestamp_t val);
+template KUZU_API void ValueVector::setValue<interval_t>(uint32_t pos, interval_t val);
+template KUZU_API void ValueVector::setValue<list_entry_t>(uint32_t pos, list_entry_t val);
 
 template<>
 void ValueVector::setValue(uint32_t pos, ku_string_t val) {
@@ -396,6 +396,10 @@ void ValueVector::setValue(uint32_t pos, ku_string_t val) {
 template<>
 void ValueVector::setValue(uint32_t pos, std::string val) {
     StringVector::addString(this, pos, val.data(), val.length());
+}
+
+void ValueVector::setNull(uint32_t pos, bool isNull) {
+    nullMask->setNull(pos, isNull);
 }
 
 void StringVector::addString(ValueVector* vector, uint32_t vectorPos, ku_string_t& srcStr) {
