+++ src/function/vector_string_functions.cpp
@@ -23,6 +23,105 @@ using namespace kuzu::common;
 namespace kuzu {
 namespace function {
 
+void BaseLowerUpperFunction::operation(common::ku_string_t& input, common::ku_string_t& result,
+    common::ValueVector& resultValueVector, bool isUpper) {
+    uint32_t resultLen = getResultLen((char*)input.getData(), input.len, isUpper);
+    result.len = resultLen;
+    if (resultLen <= common::ku_string_t::SHORT_STR_LENGTH) {
+        convertCase((char*)result.prefix, input.len, (char*)input.getData(), isUpper);
+    } else {
+        result.overflowPtr = reinterpret_cast<uint64_t>(
+            common::StringVector::getInMemOverflowBuffer(&resultValueVector)
+                ->allocateSpace(result.len));
+        auto buffer = reinterpret_cast<char*>(result.overflowPtr);
+        convertCase(buffer, input.len, (char*)input.getData(), isUpper);
+        memcpy(result.prefix, buffer, common::ku_string_t::PREFIX_LENGTH);
+    }
+}
+
+void BaseStrOperation::operation(common::ku_string_t& input, common::ku_string_t& result,
+    common::ValueVector& resultValueVector, uint32_t (*strOperation)(char* data, uint32_t len)) {
+    if (input.len <= common::ku_string_t::SHORT_STR_LENGTH) {
+        memcpy(result.prefix, input.prefix, input.len);
+        result.len = strOperation((char*)result.prefix, input.len);
+    } else {
+        result.overflowPtr = reinterpret_cast<uint64_t>(
+            common::StringVector::getInMemOverflowBuffer(&resultValueVector)
+                ->allocateSpace(input.len));
+        auto buffer = reinterpret_cast<char*>(result.overflowPtr);
+        memcpy(buffer, input.getData(), input.len);
+        result.len = strOperation(buffer, input.len);
+        memcpy(result.prefix, buffer,
+            result.len < common::ku_string_t::PREFIX_LENGTH ? result.len :
+                                                              common::ku_string_t::PREFIX_LENGTH);
+    }
+}
+
+void Concat::concat(const char* left, uint32_t leftLen, const char* right, uint32_t rightLen,
+    common::ku_string_t& result, common::ValueVector& resultValueVector) {
+    auto len = leftLen + rightLen;
+    if (len <= common::ku_string_t::SHORT_STR_LENGTH /* concat's result is short */) {
+        memcpy(result.prefix, left, leftLen);
+        memcpy(result.prefix + leftLen, right, rightLen);
+    } else {
+        result.overflowPtr = reinterpret_cast<uint64_t>(
+            common::StringVector::getInMemOverflowBuffer(&resultValueVector)->allocateSpace(len));
+        auto buffer = reinterpret_cast<char*>(result.overflowPtr);
+        memcpy(buffer, left, leftLen);
+        memcpy(buffer + leftLen, right, rightLen);
+        memcpy(result.prefix, buffer, common::ku_string_t::PREFIX_LENGTH);
+    }
+    result.len = len;
+}
+
+void Repeat::operation(common::ku_string_t& left, int64_t& right, common::ku_string_t& result,
+    common::ValueVector& resultValueVector) {
+    result.len = left.len * right;
+    if (result.len <= common::ku_string_t::SHORT_STR_LENGTH) {
+        repeatStr((char*)result.prefix, left.getAsString(), right);
+    } else {
+        result.overflowPtr = reinterpret_cast<uint64_t>(
+            common::StringVector::getInMemOverflowBuffer(&resultValueVector)
+                ->allocateSpace(result.len));
+        auto buffer = reinterpret_cast<char*>(result.overflowPtr);
+        repeatStr(buffer, left.getAsString(), right);
+        memcpy(result.prefix, buffer, common::ku_string_t::PREFIX_LENGTH);
+    }
+}
+
+void Reverse::operation(common::ku_string_t& input, common::ku_string_t& result,
+    common::ValueVector& resultValueVector) {
+    bool isAscii = true;
+    std::string inputStr = input.getAsString();
+    for (uint32_t i = 0; i < input.len; i++) {
+        if (inputStr[i] & 0x80) {
+            isAscii = false;
+            break;
+        }
+    }
+    if (isAscii) {
+        BaseStrOperation::operation(input, result, resultValueVector, reverseStr);
+    } else {
+        result.len = input.len;
+        if (result.len > common::ku_string_t::SHORT_STR_LENGTH) {
+            result.overflowPtr = reinterpret_cast<uint64_t>(
+                common::StringVector::getInMemOverflowBuffer(&resultValueVector)
+                    ->allocateSpace(input.len));
+        }
+        auto resultBuffer = result.len <= common::ku_string_t::SHORT_STR_LENGTH ?
+                                reinterpret_cast<char*>(result.prefix) :
+                                reinterpret_cast<char*>(result.overflowPtr);
+        utf8proc::utf8proc_grapheme_callback(
+            inputStr.c_str(), input.len, [&](size_t start, size_t end) {
+                memcpy(resultBuffer + input.len - end, input.getData() + start, end - start);
+                return true;
+            });
+        if (result.len > common::ku_string_t::SHORT_STR_LENGTH) {
+            memcpy(result.prefix, resultBuffer, common::ku_string_t::PREFIX_LENGTH);
+        }
+    }
+}
+
 vector_function_definitions ArrayExtractVectorFunction::getDefinitions() {
     vector_function_definitions definitions;
     definitions.emplace_back(make_unique<VectorFunctionDefinition>(ARRAY_EXTRACT_FUNC_NAME,
