+++ test/c_api/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_kuzu_test(c_api_test
+add_kuzu_api_test(c_api_test
         connection_test.cpp
         data_type_test.cpp
         database_test.cpp
