+++ test/graph_test/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_library(
         graph_test
         OBJECT
+        base_graph_test.cpp
         graph_test.cpp)
 
 target_include_directories(
@@ -11,3 +12,18 @@ target_include_directories(
 )
 
 target_link_libraries(graph_test PUBLIC GTest::GTest kuzu)
+
+
+add_library(
+        api_graph_test
+        OBJECT
+        base_graph_test.cpp)
+
+target_include_directories(
+        api_graph_test
+        PUBLIC
+        ../include/
+        ../../src/include
+)
+
+target_link_libraries(api_graph_test PUBLIC GTest::GTest kuzu_shared)
