+++ test/graph_test/graph_test.cpp
@@ -17,7 +17,7 @@ using namespace kuzu::transaction;
 namespace kuzu {
 namespace testing {
 
-void BaseGraphTest::validateColumnFilesExistence(
+void PrivateGraphTest::validateColumnFilesExistence(
     std::string fileName, bool existence, bool hasOverflow) {
     ASSERT_EQ(FileUtils::fileOrPathExists(fileName), existence);
     if (hasOverflow) {
@@ -26,7 +26,7 @@ void BaseGraphTest::validateColumnFilesExistence(
     }
 }
 
-void BaseGraphTest::validateListFilesExistence(
+void PrivateGraphTest::validateListFilesExistence(
     std::string fileName, bool existence, bool hasOverflow, bool hasHeader) {
     ASSERT_EQ(FileUtils::fileOrPathExists(fileName), existence);
     ASSERT_EQ(FileUtils::fileOrPathExists(StorageUtils::getListMetadataFName(fileName)), existence);
@@ -40,7 +40,7 @@ void BaseGraphTest::validateListFilesExistence(
     }
 }
 
-void BaseGraphTest::validateNodeColumnFilesExistence(
+void PrivateGraphTest::validateNodeColumnFilesExistence(
     NodeTableSchema* nodeTableSchema, DBFileType dbFileType, bool existence) {
     validateColumnFilesExistence(
         StorageUtils::getNodeIndexFName(databasePath, nodeTableSchema->tableID, dbFileType),
@@ -48,7 +48,7 @@ void BaseGraphTest::validateNodeColumnFilesExistence(
         containsOverflowFile(nodeTableSchema->getPrimaryKey()->getDataType()->getLogicalTypeID()));
 }
 
-void BaseGraphTest::validateRelColumnAndListFilesExistence(
+void PrivateGraphTest::validateRelColumnAndListFilesExistence(
     RelTableSchema* relTableSchema, DBFileType dbFileType, bool existence) {
     for (auto relDirection : RelDataDirectionUtils::getRelDataDirections()) {
         if (relTableSchema->getRelMultiplicity() == RelMultiplicity::MANY_ONE ||
@@ -69,7 +69,7 @@ void BaseGraphTest::validateRelColumnAndListFilesExistence(
     }
 }
 
-void BaseGraphTest::validateQueryBestPlanJoinOrder(
+void PrivateGraphTest::validateQueryBestPlanJoinOrder(
     std::string query, std::string expectedJoinOrder) {
     auto catalog = getCatalog(*database);
     auto statement = parser::Parser::parseQuery(query);
@@ -82,17 +82,7 @@ void BaseGraphTest::validateQueryBestPlanJoinOrder(
     ASSERT_STREQ(LogicalPlanUtil::encodeJoin(*plan).c_str(), expectedJoinOrder.c_str());
 }
 
-void BaseGraphTest::commitOrRollbackConnectionAndInitDBIfNecessary(
-    bool isCommit, TransactionTestType transactionTestType) {
-    commitOrRollbackConnection(isCommit, transactionTestType);
-    if (transactionTestType == TransactionTestType::RECOVERY) {
-        // This creates a new database/conn/readConn and should run the recovery algorithm.
-        createDBAndConn();
-        conn->query("BEGIN TRANSACTION");
-    }
-}
-
-void BaseGraphTest::validateRelPropertyFiles(catalog::RelTableSchema* relTableSchema,
+void PrivateGraphTest::validateRelPropertyFiles(catalog::RelTableSchema* relTableSchema,
     RelDataDirection relDirection, bool isColumnProperty, DBFileType dbFileType, bool existence) {
     for (auto& property : relTableSchema->properties) {
         auto hasOverflow = containsOverflowFile(property->getDataType()->getLogicalTypeID());
@@ -110,59 +100,7 @@ void BaseGraphTest::validateRelPropertyFiles(catalog::RelTableSchema* relTableSc
     }
 }
 
-void TestHelper::executeScript(const std::string& cypherScript, Connection& conn) {
-    std::cout << "cypherScript: " << cypherScript << std::endl;
-    if (!FileUtils::fileOrPathExists(cypherScript)) {
-        std::cout << "CpyherScript: " << cypherScript << " doesn't exist. Skipping..." << std::endl;
-        return;
-    }
-    std::ifstream file(cypherScript);
-    if (!file.is_open()) {
-        throw Exception(
-            StringUtils::string_format("Error opening file: {}, errno: {}.", cypherScript, errno));
-    }
-    std::string line;
-    while (getline(file, line)) {
-        // If this is a COPY statement, we need to append the KUZU_ROOT_DIRECTORY to the csv
-        // file path. There maybe multiple csv files in the line, so we need to find all of them.
-        std::vector<std::string> csvFilePaths;
-        size_t index = 0;
-        while (true) {
-            size_t start = line.find('"', index);
-            if (start == std::string::npos) {
-                break;
-            }
-            size_t end = line.find('"', start + 1);
-            if (end == std::string::npos) {
-                // No matching double quote, must not be a file path.
-                break;
-            }
-            std::string substr = line.substr(start + 1, end - start - 1);
-            // convert to lower case
-            auto substrLower = substr;
-            std::transform(substrLower.begin(), substrLower.end(), substrLower.begin(), ::tolower);
-            if (substrLower.find(".csv") != std::string::npos ||
-                substrLower.find(".parquet") != std::string::npos ||
-                substrLower.find(".npy") != std::string::npos ||
-                substrLower.find(".ttl") != std::string::npos) {
-                csvFilePaths.push_back(substr);
-            }
-            index = end + 1;
-        }
-        for (auto& csvFilePath : csvFilePaths) {
-            auto fullPath = appendKuzuRootPath(csvFilePath);
-            line.replace(line.find(csvFilePath), csvFilePath.length(), fullPath);
-        }
-        std::cout << "Starting to execute query: " << line << std::endl;
-        auto result = conn.query(line);
-        std::cout << "Executed query: " << line << std::endl;
-        if (!result->isSuccess()) {
-            throw Exception(StringUtils::string_format(
-                "Failed to execute statement: {}.\nError: {}", line, result->getErrorMessage()));
-        }
-    }
-}
-void BaseGraphTest::createDB(uint64_t checkpointWaitTimeout) {
+void DBTest::createDB(uint64_t checkpointWaitTimeout) {
     if (database != nullptr) {
         database.reset();
     }
@@ -170,59 +108,6 @@ void BaseGraphTest::createDB(uint64_t checkpointWaitTimeout) {
     getTransactionManager(*database)->setCheckPointWaitTimeoutForTransactionsToLeaveInMicros(
         checkpointWaitTimeout /* 10ms */);
     spdlog::set_level(spdlog::level::info);
-}
-
-void BaseGraphTest::createConns(const std::set<std::string>& connNames) {
-    if (connNames.size() == 0) { // impart a default connName
-        conn = std::make_unique<main::Connection>(database.get());
-    } else {
-        for (auto connName : connNames) {
-            if (connMap[connName] != nullptr) {
-                connMap[connName].reset();
-            }
-            connMap[connName] = std::make_unique<main::Connection>(database.get());
-        }
-    }
-}
-
-void BaseGraphTest::createDBAndConn() {
-    if (database != nullptr) {
-        database.reset();
-    }
-    database = std::make_unique<main::Database>(databasePath, *systemConfig);
-    conn = std::make_unique<main::Connection>(database.get());
-    spdlog::set_level(spdlog::level::info);
-}
-
-void BaseGraphTest::initGraph() {
-    if (conn) { // normal conn
-        TestHelper::executeScript(getInputDir() + TestHelper::SCHEMA_FILE_NAME, *conn);
-        TestHelper::executeScript(getInputDir() + TestHelper::COPY_FILE_NAME, *conn);
-    } else {
-        // choose a conn from connMap
-        TestHelper::executeScript(
-            getInputDir() + TestHelper::SCHEMA_FILE_NAME, *(connMap.begin()->second));
-        TestHelper::executeScript(
-            getInputDir() + TestHelper::COPY_FILE_NAME, *(connMap.begin()->second));
-    }
-}
-
-void BaseGraphTest::commitOrRollbackConnection(
-    bool isCommit, TransactionTestType transactionTestType) const {
-    if (transactionTestType == TransactionTestType::NORMAL_EXECUTION) {
-        if (isCommit) {
-            conn->query("COMMIT");
-        } else {
-            conn->query("ROLLBACK");
-        }
-        conn->query("BEGIN TRANSACTION");
-    } else {
-        if (isCommit) {
-            conn->query("COMMIT_SKIP_CHECKPOINT");
-        } else {
-            conn->query("ROLLBACK_SKIP_CHECKPOINT");
-        }
-    }
 }
 
 } // namespace testing
