+++ test/test_helper/CMakeLists.txt
@@ -12,3 +12,18 @@ target_include_directories(
 )
 
 target_link_libraries(test_helper PUBLIC kuzu)
+
+add_library(
+        api_test_helper
+        OBJECT
+        test_helper.cpp
+)
+
+target_include_directories(
+        api_test_helper
+        PUBLIC
+        ../include/
+        ../../src/include
+)
+
+target_link_libraries(api_test_helper PUBLIC kuzu_shared)
